refactor(#792): split project_ops_persistence_tests into base + marker/tone suite (sweep)
project_ops_persistence_tests.rs 636 -> ~342, + project_ops_persistence_more_tests.rs
~300 (both under cap; off debt). The marker/tone-block persistence tests move to the
sibling; chain_for/chain_descriptions/new_session_at become pub(super), imported via
super. adapter-gui lib tests (540) pass, zero warnings.
